+++ pygfunction/examples/load_aggregation.py
@@ -1,15 +1,10 @@
 # -*- coding: utf-8 -*-
-""" Comparison of the accuracy and computational speed of different load
-    aggregation algorithms.
+""" Example of simulation of a geothermal system.
 
     The g-function of a single borehole is calculated for boundary condition of
     uniform borehole wall temperature along the borehole. Then, the borehole
     wall temperature variations resulting from a time-varying load profile
-    are simulated using the aggregation methods of Bernier et al. (2004),
-    Liu (2005), and Claesson and Javed (2012). Results are compared to the
-    exact solution obtained by convolution in the Fourier domain.
-
-    Default parameters are used for each of the aggregation schemes.
+    are simulated using the aggregation method of Claesson and Javed (2012).
 
 """
 from __future__ import division, print_function, absolute_import
@@ -17,18 +12,9 @@ from __future__ import division, print_function, absolute_import
 import matplotlib.pyplot as plt
 from matplotlib.ticker import AutoMinorLocator
 import numpy as np
-import os
 from scipy.constants import pi
 from scipy.interpolate import interp1d
 from scipy.signal import fftconvolve
-import sys
-import time as tim
-
-# Add path to pygfunction to Python path
-packagePath = os.path.normpath(
-        os.path.join(os.path.normpath(os.path.dirname(__file__)),
-                     '..'))
-sys.path.append(packagePath)
 
 import pygfunction as gt
 
@@ -56,13 +42,8 @@ def main():
     tmax = 20.*8760. * 3600.    # Maximum time (s)
     Nt = int(np.ceil(tmax/dt))  # Number of time steps
 
-    # Load aggregation schemes
-    ClaessonJaved = gt.load_aggregation.ClaessonJaved(dt, tmax)
-    MLAA = gt.load_aggregation.MLAA(dt, tmax)
-    Liu = gt.load_aggregation.Liu(dt, tmax)
-    LoadAggSchemes = [ClaessonJaved, MLAA, Liu]
-    loadAgg_labels = ['Claesson and Javed', 'MLAA', 'Liu']
-    loadAgg_lines = ['b-', 'k--', 'r-.']
+    # Load aggregation scheme
+    LoadAgg = gt.load_aggregation.ClaessonJaved(dt, tmax)
 
     # -------------------------------------------------------------------------
     # Calculate g-function
@@ -71,49 +52,36 @@ def main():
     # The field contains only one borehole
     boreField = [gt.boreholes.Borehole(H, D, r_b, x=0., y=0.)]
     # Get time values needed for g-function evaluation
-    time_gFunc = ClaessonJaved.get_times_for_simulation()
+    time_req = LoadAgg.get_times_for_simulation()
     # Calculate g-function
-    print('Calculation of the g-function ...')
-    gFunc = gt.gfunction.uniform_temperature(boreField, time_gFunc, alpha,
+    gFunc = gt.gfunction.uniform_temperature(boreField, time_req, alpha,
                                              nSegments=nSegments)
+    # Initialize load aggregation scheme
+    LoadAgg.initialize(gFunc/(2*pi*k_s))
 
     # -------------------------------------------------------------------------
     # Simulation
     # -------------------------------------------------------------------------
-    nLoadAgg = len(LoadAggSchemes)
-    T_b = np.zeros((nLoadAgg, Nt))
+
+    time = 0.
+    i = -1
+    T_b = np.zeros(Nt)
     Q = np.zeros(Nt)
-    t_calc = np.zeros(nLoadAgg)
-    for n in range(nLoadAgg):
-        print('Simulation using {} ...'.format(loadAgg_labels[n]))
-        # Select aggregation scheme
-        LoadAgg = LoadAggSchemes[n]
-        # Interpolate g-function at require times
-        time_req = LoadAgg.get_times_for_simulation()
-        gFunc_int = interp1d(time_gFunc, gFunc)(time_req)
-        # Initialize load aggregation scheme
-        LoadAgg.initialize(np.reshape(gFunc_int, (1, 1, -1))/(2*pi*k_s))
+    while time < tmax:
+        # Increment time step by (1)
+        time += dt
+        i += 1
+        LoadAgg.next_time_step(time)
 
-        tic = tim.time()
-        time = 0.
-        i = -1
-        while time < tmax:
-            # Increment time step by (1)
-            time += dt
-            i += 1
-            LoadAgg.next_time_step(time)
+        # Evaluate heat extraction rate
+        Q[i] = synthetic_load(time/3600.)
 
-            # Evaluate heat extraction rate
-            Q[i] = synthetic_load(time/3600.)
+        # Apply current load
+        LoadAgg.set_current_load(Q[i]/H)
 
-            # Apply current load
-            LoadAgg.set_current_load(Q[i]/H)
-
-            # Evaluate borehole wall temeprature
-            deltaT_b = LoadAgg.temporal_superposition()
-            T_b[n,i] = T_g - deltaT_b
-        toc = tim.time()
-        t_calc[n] = toc - tic
+        # Evaluate borehole wall temeprature
+        deltaT_b = LoadAgg.temporal_superposition()
+        T_b[i] = T_g - deltaT_b
 
     # -------------------------------------------------------------------------
     # Calculate exact solution from convolution in the Fourier domain
@@ -124,7 +92,7 @@ def main():
     dQ[0] = Q[0]
     # Interpolated g-function
     time = np.array([(j+1)*dt for j in range(Nt)])
-    g = interp1d(time_gFunc, gFunc)(time)
+    g = interp1d(time_req, gFunc)(time)
     for i in range(1, Nt):
         dQ[i] = Q[i] - Q[i-1]
 
@@ -149,19 +117,14 @@ def main():
     # Axis labels
     ax2.set_xlabel(r'$t$ (hours)')
     ax2.set_ylabel(r'$T_b$ (degC)')
-    for n in range(nLoadAgg):
-        ax2.plot(hours, T_b[n,:],
-                 loadAgg_lines[n], lw=1.5, label=loadAgg_labels[n])
-    ax2.plot(hours, T_b_exact, 'k.', lw=1.5, label='exact')
-    ax2.legend()
+    ax2.plot(hours, T_b, 'b-', lw=1.5)
+    ax2.plot(hours, T_b_exact, 'k.', lw=1.5)
 
     ax3 = fig.add_subplot(313)
     # Axis labels
     ax3.set_xlabel(r'$t$ (hours)')
     ax3.set_ylabel(r'Error (degC)')
-    for n in range(nLoadAgg):
-        ax3.plot(hours, T_b[n,:] - T_b_exact,
-                 loadAgg_lines[n], lw=1.5, label=loadAgg_labels[n])
+    ax3.plot(hours, T_b - T_b_exact, 'b-', lw=1.5)
     # Show minor ticks
     ax1.xaxis.set_minor_locator(AutoMinorLocator())
     ax1.yaxis.set_minor_locator(AutoMinorLocator())
@@ -171,25 +134,6 @@ def main():
     ax3.yaxis.set_minor_locator(AutoMinorLocator())
     # Adjust to plot window
     plt.tight_layout()
-
-    # -------------------------------------------------------------------------
-    # Print performance metrics
-    # -------------------------------------------------------------------------
-
-    # Maximum errors in evaluation of borehole wall temperatures
-    maxError = np.array([np.max(np.abs(T_b[n,:]-T_b_exact))
-                         for n in range(nLoadAgg)])
-    # Print results
-    print('Simulation results')
-    horizontalLine = '-'*66
-    for n in range(nLoadAgg):
-        print(horizontalLine)
-        print(loadAgg_labels[n])
-        print()
-        print('Maximum absolute error : {} degC'.format(maxError[n]))
-        print('Calculation time : {} seconds'.format(t_calc[n]))
-        print()
-    print(horizontalLine)
 
     return
 
@@ -209,15 +153,14 @@ def synthetic_load(x):
     G = 0.95
 
     func = (168.0-C)/168.0
-    for i in [1, 2, 3]:
-        func += 1.0/(i*pi)*(np.cos(C*pi*i/84.0) - 1.0) \
-                          *(np.sin(pi*i/84.0*(x - B)))
-    func = func*A*np.sin(pi/12.0*(x - B)) \
-        *np.sin(pi/4380.0*(x - B))
+    for i in [1,2,3]:
+        func += 1.0/(i*pi)*(np.cos(C*pi*i/84.0)-1.0) \
+                          *(np.sin(pi*i/84.0*(x-B)))
+    func = func*A*np.sin(pi/12.0*(x-B)) \
+           *np.sin(pi/4380.0*(x-B))
 
-    y = func + (-1.0)**np.floor(D/8760.0*(x - B))*abs(func) \
-        + E*(-1.0)**np.floor(D/8760.0*(x - B)) \
-        /np.sign(np.cos(D*pi/4380.0*(x - F)) + G)
+    y = func + (-1.0)**np.floor(D/8760.0*(x-B))*abs(func) \
+      + E*(-1.0)**np.floor(D/8760.0*(x-B))/np.sign(np.cos(D*pi/4380.0*(x-F))+G)
     return -np.array([y])